docs: rework docs content, Star Trek examples, IDE-style code blocks
Reworks the docs site per review feedback:

- Star Trek-theme every example spec (hero, examples, getting-started)
- Make the hero snippet a data-driven spec and drop the "kotlin" language badge
- Remove the Sneak Preview page (title page now covers that content)
- Replace How It Works with a focused Examples page (5 short specs
  covering docs-as-specs, natural assertions, data tables, fixtures,
  and experimental extension support)
- Drop the "independent add-on" footer disclaimer
- Introduce CodeWindow, an IDE-style code block (title bar, shading,
  line reveal-on-scroll) replacing RevealCode/TerminalWindow, and
  rename the carousel components for ESLint's multi-word-component rule
- Unify typography: one code font size and one meta/caption text size as
  CSS custom properties, switch the code font to JetBrains Mono
- Fix CodeWindow's code becoming center-aligned per-line, and force
  minmax(0, 1fr) grid tracks, when used outside .vp-doc (e.g. the home
  page hero) where VitePress's own text-align/display overrides don't
  reach, and where an unbreakable list item can otherwise blow out a
  bare 1fr track's width on mobile
