+++ spockk-docs/docs/.vitepress/theme/index.ts
@@ -15,21 +15,23 @@
 import type { Theme } from 'vitepress'
 import DefaultTheme from 'vitepress/theme'
 
+import '@fontsource/jetbrains-mono/400.css'
+import '@fontsource/jetbrains-mono/500.css'
+import '@fontsource/jetbrains-mono/700.css'
+
+import CodeCarousel from './components/CodeCarousel.vue'
+import CarouselSlide from './components/CarouselSlide.vue'
+import CodeWindow from './components/CodeWindow.vue'
 import ComparisonPanels from './components/ComparisonPanels.vue'
-import RevealCode from './components/RevealCode.vue'
-import TerminalWindow from './components/TerminalWindow.vue'
-import TransformCarousel from './components/TransformCarousel.vue'
-import TransformSlide from './components/TransformSlide.vue'
 import { vReveal } from './reveal'
 import './style.css'
 
 export default {
   extends: DefaultTheme,
   enhanceApp({ app }) {
-    app.component('RevealCode', RevealCode)
-    app.component('TerminalWindow', TerminalWindow)
-    app.component('TransformCarousel', TransformCarousel)
-    app.component('TransformSlide', TransformSlide)
+    app.component('CodeWindow', CodeWindow)
+    app.component('CodeCarousel', CodeCarousel)
+    app.component('CarouselSlide', CarouselSlide)
     app.component('ComparisonPanels', ComparisonPanels)
     app.directive('reveal', vReveal)
   },
